Moves heathHelpers.tst.mjs to Playwright
